Add file stress-tests/geometry/utilities.h for things in multiple stress-test
